Fix(backend): [SCRUM-95] Delete ChecklistItems controller, move ChecklistItem HTTP methods into user controller
