+++ HelloCity.Api/Controllers/UserController.cs
@@ -5,6 +5,7 @@ using Microsoft.AspNetCore.Mvc;
 using HelloCity.Models.Entities;
 using AutoMapper;
 using Microsoft.JSInterop.Infrastructure;
+using HelloCity.Api.DTOs.ChecklistItem;
 
 namespace HelloCity.Api.Controllers
 {
@@ -15,9 +16,10 @@ namespace HelloCity.Api.Controllers
         private readonly IUserService _userService;
         private readonly IMapper _mapper;
         private readonly ILogger<UserController> _logger;
-
-        public UserController(IUserService userService, IMapper mapper, ILogger<UserController> logger)
+        private readonly IChecklistItemService _checklistItemService;
+        public UserController(IUserService userService, IMapper mapper, ILogger<UserController> logger, IChecklistItemService checklistItemService)
         {
+            _checklistItemService = checklistItemService;
             _userService = userService;
             _mapper = mapper;
             _logger = logger;
@@ -66,7 +68,7 @@ namespace HelloCity.Api.Controllers
 
             return CreatedAtAction(
                 nameof(GetUserProfile),
-                new { id = userDto.UserId }, 
+                new { id = userDto.UserId },
                 new
                 {
                     message = "create user successfully",
@@ -105,6 +107,43 @@ namespace HelloCity.Api.Controllers
                 }
             });
         }
-
+        [HttpPost("{userId}/checklist-item")]
+        public async Task<ActionResult<ChecklistItem>> CreateChecklistItem(Guid userId, CreateChecklistItemDto newChecklistItemDto)
+        {
+            _logger.LogInformation("Creating checklist for User with ID: {userId}", userId);
+            try
+            {
+                var newChecklistItem = _mapper.Map<ChecklistItem>(newChecklistItemDto);
+                var savedChecklistItem = await _checklistItemService.AddChecklistItemAsync(userId, newChecklistItem);
+                var checklistItemDto = _mapper.Map<ChecklistItemDto>(savedChecklistItem);
+                return Created("/api/checklist-item/" + checklistItemDto.ChecklistItemId, checklistItemDto);
+            }
+            catch (KeyNotFoundException ex)
+            {
+                _logger.LogWarning(ex, "User not found with ID: {userId}", userId);
+                return NotFound("User not found");
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex, "Unexpected error while creating checklist item for User with ID: {userId}", userId);
+                return Problem("An unexpected error occurred.");
+            }
+        }
+        [HttpGet("{userId}/checklist-item")]
+        public async Task<ActionResult<List<ChecklistItem>>> GetChecklistItems(Guid userId)
+        {
+            _logger.LogInformation("Getting checklist for User with ID: {userId}", userId);
+            try
+            {
+                var userChecklistItems = await _checklistItemService.GetChecklistItemsAsync(userId);
+                var checklistItemDtos = _mapper.Map<List<ChecklistItemDto>>(userChecklistItems);
+                return Ok(checklistItemDtos);
+            }
+            catch (KeyNotFoundException ex)
+            {
+                _logger.LogWarning(ex, "User not found with ID: {userId}", userId);
+                return NotFound("User not found");
+            }
+        }
     }
 }
