+++ HelloCity.Repository/ChecklistItemRepository.cs
@@ -16,7 +16,7 @@ namespace HelloCity.Repository{
       var user = await _context.Users
         .Include(u => u.ChecklistItems) 
         .FirstOrDefaultAsync(u => u.UserId == id);
-      if (user == null) return null;
+      if (user is null) return null;
       return user.ChecklistItems;
     }
     public async Task<ChecklistItem> AddChecklistItemAsync(Guid id, ChecklistItem newChecklistItem)
